+++ tests/test_models/test_timestamped_model.py
@@ -15,6 +15,13 @@ class TimeStampedModelTests(TestCase):
             t1 = TimeStamp.objects.create()
         self.assertEqual(t1.created, datetime(2016, 1, 1))
 
+    def test_created_sets_modified(self):
+        '''
+        Ensure that on creation that modifed is set exactly equal to created.
+        '''
+        t1 = TimeStamp.objects.create()
+        self.assertEqual(t1.created, t1.modified)
+
     def test_modified(self):
         with freeze_time(datetime(2016, 1, 1)):
             t1 = TimeStamp.objects.create()